refactor: relocate back button from title bar to settings sidebar
Move the back navigation button from the title bar to the settings sidebar footer for better UX and cleaner layout.
